removed errors and bugs. Changed path problem from physical layer.

+++ schema/destroydb.c
@@ -13,17 +13,17 @@
  * deletes an existing file
  *
  *  file_name: name of the file that has to be deleted
- *  db_name: name of the database
+ *  db_path: name and address of the database
  *  
  *  returns: OK upon successful deletion of file.
  *           NOTOK: file does not exist
  */
 
-int delete_file(char* file_name, char* db_name)
+int delete_file(char* file_name, char* db_path)
 {
-    char full_file_path[3 * RELNAME];
+    char full_file_path[MAXPATH + RELNAME];
 
-    sprintf(full_file_path,"%s/%s/%s",PATH,db_name,file_name);
+    sprintf(full_file_path,"%s/%s",db_path,file_name);
 
     if(remove(full_file_path) == 0){
         return OK;
@@ -31,7 +31,6 @@ int delete_file(char* file_name, char* db_name)
     else{
         return NOTOK;
     } /* File does not exist*/
-
 }
 
 /*
@@ -39,23 +38,20 @@ int delete_file(char* file_name, char* db_name)
  * ------------------------
  * deletes an existing directory
  *
- *  dir_name: name of the directory that has to be deleted
+ *  dir_path: name and address of the directory that has to be deleted
  *  
  *  returns: OK    upon successful deletion of directory.
  *           NOTOK directory does not exist
  */
 
-int delete_dir(char* dir_name)
+int delete_dir(char* dir_path)
 {
     DIR *dir_handler;
     struct dirent *file_handler;
 
-    char dir_path[2 * RELNAME + 2];
-    sprintf(dir_path,"%s/%s",PATH,dir_name);
-
     if( (dir_handler = opendir(dir_path)) != NULL ){
         while ((file_handler = readdir (dir_handler)) != NULL) {
-            delete_file(file_handler->d_name, dir_name);
+            delete_file(file_handler->d_name, dir_path);
         }
         closedir (dir_handler);
         rmdir(dir_path);
@@ -83,6 +79,9 @@ int DestroyDB(int argc,char **argv)
 {
     if(argc < 2)
         return ErrorMsgs(ARGC_INSUFFICIENT,g_print_flag);
+
+    if(strcmp(g_db_name,"") != 0)
+        return ErrorMsgs(DB_NOT_CLOSED, g_print_flag);
 
     if(delete_dir(argv[1]) == NOTOK)
         return ErrorMsgs(DBNAME_INVALID,g_print_flag);
